Remove restart lnet handler
This is being removed as in practice it doesn't work. The problem is
that you can't just restart LNET on a server since the modules will be
in use by any active targets that are mounted.

This means that we can't restart LNET on configuration change - that
kind of action will have to be a separate on-shot playbook run, or by
running the role with state: absent to shutdown Lustre fully, and then
state: present to start it up again fully.